+++ src/views/admin-cms-content-property/_form.php
@@ -109,13 +109,31 @@ JS
         ]
     ); ?>
 
+
+
+<? endif; ?>
+
+<? if ($tree_id = \Yii::$app->request->get('tree_id')) : ?>
+
+    <div style="display: none;">
+    <?= $form->field($model, 'cmsTrees')->checkboxList(
+        \yii\helpers\ArrayHelper::map(
+            \skeeks\cms\models\CmsTree::find()->where(['id' => $tree_id])->all(),
+            'id',
+            'name'
+        ),
+        [
+            'value' => [$tree_id]
+        ]
+    ); ?>
+    </div>
+<? else : ?>
     <?= $form->field($model, 'cmsTrees')->widget(
         \skeeks\cms\widgets\formInputs\selectTree\SelectTreeInputWidget::class,
         [
             'multiple' => true,
         ]
     ); ?>
-
 <? endif; ?>
 
 <?= $form->fieldSetEnd(); ?>